refactor: replace multiple asserts with assertAll
