+++ src/main/java/racing/domain/RacingCarGame.java
@@ -1,6 +1,9 @@
 package racing.domain;
 
 import racing.config.GameConfig;
+import racing.domain.model.Car;
+import racing.domain.model.Cars;
+import racing.domain.model.Round;
 
 import java.util.*;
