+++ src/main/java/racing/domain/model/Car.java
@@ -1,4 +1,4 @@
-package racing.domain;
+package racing.domain.model;
 
 public class Car {
