+++ src/test/java/racing/domain/CarTest.java
@@ -2,6 +2,7 @@ package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
